Move ImageRequest back into img.py
